refactor: hooks 폴더구조 변경
- business 로직을 담은 hook은 domain
- UI util의 재활용에 관련된 hook은 UI로 이동
